Refactor sensor architecture to use SensorType enum
Introduced the `SensorType` enum to standardize sensor type definitions. Updated all relevant classes, including `Sensor`, `SensorFactory`, and `SimulationServiceImpl`, to replace string-based type handling with the new enum. Enhanced data consistency, maintainability, and type safety across the system.

+++ src/main/java/pl/edu/pjwstk/s25819/smartcity/sensors/service/SensorService.java
@@ -13,7 +13,7 @@ public interface SensorService {
 
     SensorResponseDto createSensor(SensorRequestDto sensorRequestDto);
 
-    SensorResponseDto updateSensor(int id, SensorChangeStatusRequestDto sensorChangeStatusRequestDto);
+    SensorResponseDto updateSensor(long id, SensorChangeStatusRequestDto sensorChangeStatusRequestDto);
 
     List<SensorTypeResponseDto> getSensorTypes();
 }
